requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-experimental>=0.0.47
langgraph>=0.0.40
google-generativeai>=0.3.0

# Tool dependencies
duckduckgo-search>=3.9.0
wikipedia-api>=0.5.4
arxiv>=2.0.0

# CLI dependencies
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0

# Additional dependencies
pydantic>=2.0.0
typing-extensions>=4.8.0

# Optional: For web interface (uncomment if needed)
# langgraph[studio]>=0.0.40
# streamlit>=1.28.0 